requirements.txt
gevent==1.1a1 # https://github.com/gevent/gevent/issues/441#issuecomment-162921780
snowplow-tracker==0.8.0
click==6.7
kafka-python==1.3.3
thriftpy==0.3.9